Add tests to aid in DCE connection debugging
