+++ src/App.css
@@ -18,3 +18,7 @@
 	color: rgb(0,0,0);
 	font-weight: 600;
 }
+
+.disclaimer {
+	font-size: small;
+}
